+++ oap-server/server-query-plugin/query-graphql-plugin/src/main/java/org/apache/skywalking/oap/query/graphql/mqe/rt/MQEVisitor.java
@@ -26,7 +26,7 @@ import java.util.List;
 import java.util.Optional;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.skywalking.oap.query.graphql.mqe.rt.exception.IllegalExpressionException;
-import org.apache.skywalking.oap.query.graphql.mqe.rt.operation.AggregateLabelOp;
+import org.apache.skywalking.oap.query.graphql.mqe.rt.operation.AggregateLabelsOp;
 import org.apache.skywalking.oap.query.graphql.mqe.rt.operation.AggregationOp;
 import org.apache.skywalking.oap.query.graphql.resolver.MetricsQuery;
 import org.apache.skywalking.oap.query.graphql.resolver.RecordsQuery;
@@ -152,8 +152,8 @@ public class MQEVisitor extends MQEParserBaseVisitor<ExpressionResult> {
     }
 
     @Override
-    public ExpressionResult visitAggregateLabelOp(final MQEParser.AggregateLabelOpContext ctx) {
-        int aggregateLabelType = ctx.aggregateLabelFunc().getStart().getType();
+    public ExpressionResult visitAggregateLabelsOp(final MQEParser.AggregateLabelsOpContext ctx) {
+        int funcType = ctx.aggregateLabelsFunc().getStart().getType();
         ExpressionResult expResult = visit(ctx.expression());
         if (StringUtil.isNotEmpty(expResult.getError())) {
             return expResult;
@@ -165,7 +165,7 @@ public class MQEVisitor extends MQEParserBaseVisitor<ExpressionResult> {
         }
 
         try {
-            return AggregateLabelOp.doAggregateLabelOp(expResult, aggregateLabelType);
+            return AggregateLabelsOp.doAggregateLabelsOp(expResult, funcType);
         } catch (IllegalExpressionException e) {
             ExpressionResult result = new ExpressionResult();
             result.setType(ExpressionResultType.UNKNOWN);