+++ oap-server/server-query-plugin/query-graphql-plugin/src/main/java/org/apache/skywalking/oap/query/graphql/mqe/rt/operation/aggregatelabels/AggregateLabelsFunc.java
@@ -16,34 +16,10 @@
  *
  */
 
-package org.apache.skywalking.oap.query.graphql.mqe.rt.operation.aggregatelabel;
+package org.apache.skywalking.oap.query.graphql.mqe.rt.operation.aggregatelabels;
 
-public class AvgAggregateLabelFunc implements AggregateLabelFunc {
+public interface AggregateLabelsFunc {
+    void combine(Double value);
 
-    private Double sum;
-    private Integer count = 0;
-
-    @Override
-    public void combine(final Double value) {
-        if (value == null) {
-            return;
-        }
-
-        if (sum == null) {
-            sum = value;
-        } else {
-            sum = sum + value;
-        }
-
-        count++;
-    }
-
-    @Override
-    public Double getResult() {
-        if (sum == null) {
-            return null;
-        }
-
-        return sum / count;
-    }
+    Double getResult();
 }