+++ oap-server/server-query-plugin/query-graphql-plugin/src/main/java/org/apache/skywalking/oap/query/graphql/mqe/rt/operation/aggregatelabels/AvgAggregateLabelsFunc.java
@@ -16,8 +16,34 @@
  *
  */
 
-package org.apache.skywalking.oap.query.graphql.mqe.rt.operation.aggregatelabel;
+package org.apache.skywalking.oap.query.graphql.mqe.rt.operation.aggregatelabels;
 
-public interface AggregateLabelFuncFactory {
-     AggregateLabelFunc getAggregateLabelFunc();
+public class AvgAggregateLabelsFunc implements AggregateLabelsFunc {
+
+    private Double sum;
+    private Integer count = 0;
+
+    @Override
+    public void combine(final Double value) {
+        if (value == null) {
+            return;
+        }
+
+        if (sum == null) {
+            sum = value;
+        } else {
+            sum = sum + value;
+        }
+
+        count++;
+    }
+
+    @Override
+    public Double getResult() {
+        if (sum == null) {
+            return null;
+        }
+
+        return sum / count;
+    }
 }