+++ docs/en/changes/changes.md
@@ -36,7 +36,7 @@
 * Accept `length=4000` parameter value of the event. It was 2000. 
 * Tolerate parameter value in illegal JSON format.
 * Update BanyanDB Java Client to 0.4.0
-* Support aggregate LabeledValue in MQE.
+* Support aggregate `Labeled Value Metrics` in MQE.
 
 #### UI